fix(tests): scope the relay-kill e2e test's bridge to the test
test_relay_child_death_respawns_and_recovers kills the shared
module-scoped bridge's relay child; the respawned child's pipe-reader
threads are created mid-test and legitimately live until module
teardown, so the per-test thread-leak monitor flags them. This mostly
went unnoticed because the monitor keys on thread idents and the new
readers usually recycle the just-joined old readers' idents, dodging
the check by luck; when scheduling breaks the reuse (seen on the arm64
CI runner) the test errors at teardown.

Give the destructive test its own function-scoped bridge so the
respawned child dies inside the test scope, and move the stop_module
helper (Module.stop + executor reap) into e2e_support for both test
files to share.
